+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+* Added support for Event Notifications via `b2_bucket_notification_rules` resource and data source
+
 ### Infrastructure
 * Use Python 3.13 for embedded pybindings